+++ packages/sdk/src/sdk/notesManager/index.ts
@@ -1,9 +1,22 @@
-import { IncludedNote } from "../note";
+import { FlaxDB } from "../db";
+import { IncludedNoteStruct } from "../note";
 
-export interface NotesManager {
-  gatherNewRefunds(): Promise<IncludedNote[]>;
-  // TODO: abstract gatherNewSpends():
-  // TODO: method to call two gather functions and update DB accordingly
+export type RefundEvent = IncludedNoteStruct;
+export interface SpendEvent {
+  oldNoteNullifier: bigint;
+  valueSpent: bigint;
+  merkleIndex: number;
+}
+
+export abstract class NotesManager {
+  db: FlaxDB;
+
+  constructor(db: FlaxDB) {
+    this.db = db;
+  }
+
+  abstract fetchAndStoreRefunds(): Promise<void>;
+  abstract fetchAndApplySpends(): Promise<void>;
 }
 
 export { LocalNotesManager } from "./local";