+++ packages/sdk/src/sdk/notesManager/local.ts
@@ -1,17 +1,20 @@
 import { Wallet, Wallet__factory } from "@flax/contracts";
-import { RefundEvent } from "@flax/contracts/dist/src/Wallet";
 import { ethers } from "ethers";
 import { Address } from "../../commonTypes";
 import { FlaxDB } from "../db";
-import { IncludedNote } from "../note";
 import { query } from "../utils";
+import {
+  RefundEvent as EthRefundEvent,
+  SpendEvent as EthSpendEvent,
+} from "@flax/contracts/dist/src/Wallet";
 import { NotesManager } from ".";
+import { IncludedNote } from "../note";
 
 const DEFAULT_START_BLOCK = 0;
 const REFUNDS_LAST_INDEXED_BLOCK = "REFUNDS_LAST_INDEXED_BLOCK";
+const SPENDS_LAST_INDEXED_BLOCK = "SPENDS_LAST_INDEXED_BLOCK";
 
-export class LocalNotesManager implements NotesManager {
-  db: FlaxDB;
+export class LocalNotesManager extends NotesManager {
   walletContract: Wallet;
   provider: ethers.providers.Provider;
 
@@ -20,20 +23,20 @@ export class LocalNotesManager implements NotesManager {
     provider: ethers.providers.Provider,
     db: FlaxDB
   ) {
-    this.db = db;
+    super(db);
     this.provider = provider;
     this.walletContract = Wallet__factory.connect(walletAddress, this.provider);
   }
 
-  async gatherNewRefunds(): Promise<IncludedNote[]> {
+  async fetchAndStoreRefunds(): Promise<void> {
     const maybeLastSeen = this.db.getKv(REFUNDS_LAST_INDEXED_BLOCK);
     const lastSeen = maybeLastSeen
-      ? parseInt(maybeLastSeen)
+      ? parseInt(maybeLastSeen) + 1
       : DEFAULT_START_BLOCK; // TODO: load default from network-specific config
     const latestBlock = await this.provider.getBlockNumber();
 
     const filter = this.walletContract.filters.Refund();
-    let events: RefundEvent[] = await query(
+    let events: EthRefundEvent[] = await query(
       this.walletContract,
       filter,
       lastSeen,
@@ -42,7 +45,7 @@ export class LocalNotesManager implements NotesManager {
 
     events = events.sort((a, b) => a.blockNumber - b.blockNumber);
 
-    return events.map((event) => {
+    const newNotes = events.map((event) => {
       const { refundAddr, nonce, asset, id, value, merkleIndex } = event.args;
       const { h1X, h1Y, h2X, h2Y } = refundAddr;
       return new IncludedNote({
@@ -59,5 +62,39 @@ export class LocalNotesManager implements NotesManager {
         merkleIndex: merkleIndex.toNumber(),
       });
     });
+
+    await this.db.storeNotes(newNotes);
+    await this.db.putKv(REFUNDS_LAST_INDEXED_BLOCK, latestBlock.toString());
+  }
+  //newNonce = H(vk, nf)
+  async fetchAndApplySpends(): Promise<void> {
+    const maybeLastSeen = this.db.getKv(SPENDS_LAST_INDEXED_BLOCK);
+    const lastSeen = maybeLastSeen
+      ? parseInt(maybeLastSeen) + 1
+      : DEFAULT_START_BLOCK; // TODO: load default from network-specific config
+    const latestBlock = await this.provider.getBlockNumber();
+
+    const filter = this.walletContract.filters.Spend();
+    let events: EthSpendEvent[] = await query(
+      this.walletContract,
+      filter,
+      lastSeen,
+      latestBlock
+    );
+
+    events = events.sort((a, b) => a.blockNumber - b.blockNumber);
+
+    const newSpends = events.map((event) => {
+      const { oldNoteNullifier, valueSpent, merkleIndex } = event.args;
+      return {
+        oldNoteNullifier: oldNoteNullifier.toBigInt(),
+        valueSpent: valueSpent.toBigInt(),
+        merkleIndex: merkleIndex.toNumber(),
+      };
+    });
+    console.log(newSpends);
+
+    // Find/remove old note and store new one by changing refundAddr and
+    // calculating new nonce
   }
 }