+++ packages/contracts/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-- Add `NewNoteFromSpend` event and emit on `handleSpend`
+- Add `Spend` event and emit on `handleSpend`
 - Rename spend transaction `value` to `valueToSpend`
 - Fix merkle index bug where `insert8` and `insert16` only incremented `numberOfLeaves` by 1
 - Start test suite for `PoseidonBatchBinaryMerkle`